requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
sqlalchemy>=1.4.0
pymysql>=1.0.0
python-dotenv>=0.19.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0
notebook>=6.4.0